chore: more improvements
