+++ packages/uniwind/src/core/web/index.ts
@@ -1,5 +1,3 @@
 export * from './cssListener'
-export * from './formatColor'
 export * from './getWebStyles'
-export * from './parseCSSValue'
 export * from './webUtils'
