--- packages/uniwind/src/core/web/parseCSSValue.ts
@@ -1,10 +1,0 @@
-import { parse } from 'culori'
-import { formatColor } from './formatColor'
-
-export const parseCSSValue = (value: string) => {
-    if (isNaN(Number(value)) && parse(value) !== undefined) {
-        return formatColor(value)
-    }
-
-    return value
-}
